Finished Hello World App
The app now has a label and some icons. Things changed a bit between iOS 6 and 7. The icons needed to be scaled
to be 58x58 and 120x120 (from 57x57 and 114x114 respectively). I’m sure there will be more differences that
will make learning with an old iOS 6 book challenging with iOS 7. Oh well...