Created a Simple UI
Created a label and some buttons. The buttons are linked to the same
action which updates the label using the title of the button. In iOS 7
this is a little different than the iOS 6 stuff the book has. The member
variable name for the label needed an underscore first. And the .m file
could not be cleared out as much as the book instructed - XCode was
unable to create actions without that interface part still in the file.